Implementing underlying machinery to support BQM parsing in plugins
